Update package version and copyright owner, refactoring some methods and documentations outer function still the same
The package version in pyproject.toml has been upgraded from 0.0.4 to 0.0.5. Additionally, the copyright owner in the LICENSE file has been updated to A. Piad .

In previous commit this was change to Jesús E. I have my own reserve about this because the original author and programmer was and still is Piad even if this is a derivative work with some changes.